fixed aritmetics in EnvpSize

+++ c/class/ws3/ws3head.h
@@ -7,7 +7,7 @@ void *StringToLower(char *);
 void *PrintEnv(const char **);
 char **CopyEnv(const char **);
 void CleanEnvCopy(char **envp);
-int EnvpSize(const char **envp); /*return Envp size - Num of pointers */
+long int EnvpSize(const char **envp); /*return Envp size - Num of pointers */
 char *MyStrDup(const char *s);
 
 #endif
